Expense type deletion error (#5952)
1) Modified relation between te expense type and te
paid expense to be bidirectional
2) Refactored expense type list page to only show the delete
option if there are no te paid expenses related to it
3) Refactored expense type delete action

+++ src/Opit/OpitHrm/TravelBundle/Controller/AdminController.php
@@ -125,7 +125,10 @@ class AdminController extends Controller
 
         foreach ($expenseTypeId as $id) {
             $expenseType = $entityManager->getRepository('OpitOpitHrmTravelBundle:TEExpenseType')->find($id);
-            $entityManager->remove($expenseType);
+
+            if (0 === count($expenseType->getTEPaidExpenses())) {
+                $entityManager->remove($expenseType);
+            }
         }
 
         $entityManager->flush();
